+++ src/main/java/yesman/epicfight/mixin/common/MixinPlayer.java
@@ -6,9 +6,11 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.Redirect;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
+import net.minecraft.client.player.LocalPlayer;
 import net.minecraft.world.damagesource.CombatTracker;
 import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.entity.player.Player;
+import yesman.epicfight.api.client.camera.EpicFightCameraAPI;
 import yesman.epicfight.world.capabilities.EpicFightCapabilities;
 import yesman.epicfight.world.capabilities.entitypatch.LivingEntityPatch;
 import yesman.epicfight.world.capabilities.entitypatch.player.PlayerPatch;
@@ -16,19 +18,23 @@ import yesman.epicfight.world.capabilities.entitypatch.player.PlayerPatch;
 @Mixin(value = Player.class)
 public abstract class MixinPlayer {
 	@Inject(at = @At(value = "TAIL"), method = "<clinit>")
-	private static void epicfight_staticInitialize(CallbackInfo callbackInfo) {
+	private static void epicfight$staticInitialize(CallbackInfo callbackInfo) {
 		PlayerPatch.initPlayerDataAccessor();
 	}
 	
 	@Inject(at = @At(value = "TAIL"), method = "defineSynchedData()V", cancellable = true)
-	protected void epicfight_defineSynchedData(CallbackInfo info) {
+	protected void epicfight$defineSynchedData(CallbackInfo info) {
 		PlayerPatch.createSyncedEntityData((Player)(Object)this);
 	}
 	
-	@Redirect(    at = @At( value = "INVOKE",
-			  target = "Lnet/minecraft/world/damagesource/CombatTracker;recordDamage(Lnet/minecraft/world/damagesource/DamageSource;F)V"),
-		      method = "actuallyHurt(Lnet/minecraft/world/damagesource/DamageSource;F)V")
-	private void epicfight_recordDamage(CombatTracker self, DamageSource damagesource, float damage) {
+	@Redirect(
+		at = @At(
+			value = "INVOKE",
+			target = "Lnet/minecraft/world/damagesource/CombatTracker;recordDamage(Lnet/minecraft/world/damagesource/DamageSource;F)V"
+		),
+		method = "actuallyHurt(Lnet/minecraft/world/damagesource/DamageSource;F)V"
+	)
+	private void epicfight$recordDamage(CombatTracker self, DamageSource damagesource, float damage) {
 		LivingEntityPatch<?> entitypatch = EpicFightCapabilities.getEntityPatch(damagesource.getEntity(), LivingEntityPatch.class);
 
 		if (entitypatch != null) {
@@ -37,4 +43,19 @@ public abstract class MixinPlayer {
 		
 		self.recordDamage(damagesource, damage);
 	}
+	
+	@Redirect(
+        at = @At(
+            value = "INVOKE",
+            target = "Lnet/minecraft/world/entity/player/Player;getYRot()F"
+        ),
+        method = "serverAiStep()V"
+    )
+    private float epicfight$serverAiStep(Player player) {
+		if (player.isLocalPlayer()) {
+			return EpicFightCameraAPI.getInstance().getYRotForHead((LocalPlayer)player);
+		}
+		
+		return player.getYRot();
+    }
 }